# storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "VoIP Termination System API",
    "version": "1.0.0",
    "description": "API REST para el sistema de terminacion VoIP. Gestiona clientes, carriers, CDRs, tarifas, reportes, fraude y mas.",
    "contact": {
      "email": "admin@tellmetelecom.com"
    }
  },
  "servers": [
    {
      "url": "https://sw1.tellmetelecom.com/api/v1",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Token"
      }
    },
    "schemas": {
      "Customer": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "uuid": {"type": "string", "format": "uuid"},
          "name": {"type": "string"},
          "company": {"type": "string"},
          "email": {"type": "string", "format": "email"},
          "phone": {"type": "string"},
          "max_channels": {"type": "integer", "default": 10},
          "max_cps": {"type": "integer", "default": 5},
          "max_daily_minutes": {"type": "integer", "nullable": true},
          "max_monthly_minutes": {"type": "integer", "nullable": true},
          "used_daily_minutes": {"type": "integer"},
          "used_monthly_minutes": {"type": "integer"},
          "active": {"type": "boolean"},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
...
    }
  },
  "paths": {
...
    "/customers": {
      "get": {
        "summary": "Listar clientes",
        "tags": ["Customers"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "active", "in": "query", "schema": {"type": "boolean"}},
          {"name": "search", "in": "query", "schema": {"type": "string"}},
          {"name": "per_page", "in": "query", "schema": {"type": "integer", "default": 50}}
        ],
        "responses": {
          "200": {"description": "Lista de clientes"}
        }
      },
      "post": {
        "summary": "Crear cliente",
        "tags": ["Customers"],
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {"type": "string"},
                  "company": {"type": "string"},
                  "email": {"type": "string"},
                  "max_channels": {"type": "integer"},
                  "max_cps": {"type": "integer"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {"description": "Cliente creado"},
          "422": {"description": "Error de validacion"}
        }
      }
    },
    "/customers/{id}": {
      "get": {
        "summary": "Obtener cliente",
        "tags": ["Customers"],
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {
          "200": {"description": "Detalle del cliente"},
          "404": {"description": "No encontrado"}
        }
      },
      "put": {
        "summary": "Actualizar cliente",
        "tags": ["Customers"],
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Customer"}}}},
        "responses": {"200": {"description": "Cliente actualizado"}}
      },
      "delete": {
        "summary": "Eliminar cliente",
        "tags": ["Customers"],
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {"200": {"description": "Eliminado"}, "409": {"description": "Tiene llamadas activas"}}
      }
    },
...
  },
...
}